+++ ressources/rolemapping_interactor.go
@@ -5,6 +5,8 @@
 package ressources
 
 import (
+	"database/sql"
+
 	"github.com/Solher/auth-scaffold/domain"
 	"github.com/Solher/auth-scaffold/interfaces"
 )
@@ -12,12 +14,13 @@ import (
 type AbstractRoleMappingRepo interface {
 	Create(rolemappings []domain.RoleMapping) ([]domain.RoleMapping, error)
 	CreateOne(rolemapping *domain.RoleMapping) (*domain.RoleMapping, error)
-	Find(filter *interfaces.Filter) ([]domain.RoleMapping, error)
-	FindByID(id int, filter *interfaces.Filter) (*domain.RoleMapping, error)
-	Upsert(rolemappings []domain.RoleMapping) ([]domain.RoleMapping, error)
-	UpsertOne(rolemapping *domain.RoleMapping) (*domain.RoleMapping, error)
-	DeleteAll(filter *interfaces.Filter) error
-	DeleteByID(id int) error
+	Find(filter *interfaces.Filter, ownerRelations []domain.Relation) ([]domain.RoleMapping, error)
+	FindByID(id int, filter *interfaces.Filter, ownerRelations []domain.Relation) (*domain.RoleMapping, error)
+	Upsert(rolemappings []domain.RoleMapping, filter *interfaces.Filter, ownerRelations []domain.Relation) ([]domain.RoleMapping, error)
+	UpsertOne(rolemapping *domain.RoleMapping, filter *interfaces.Filter, ownerRelations []domain.Relation) (*domain.RoleMapping, error)
+	DeleteAll(filter *interfaces.Filter, ownerRelations []domain.Relation) error
+	DeleteByID(id int, filter *interfaces.Filter, ownerRelations []domain.Relation) error
+	Raw(query string, values ...interface{}) (*sql.Rows, error)
 }
 
 type RoleMappingInter struct {
@@ -38,32 +41,32 @@ func (i *RoleMappingInter) CreateOne(rolemapping *domain.RoleMapping) (*domain.R
 	return rolemapping, err
 }
 
-func (i *RoleMappingInter) Find(filter *interfaces.Filter) ([]domain.RoleMapping, error) {
-	rolemappings, err := i.repo.Find(filter)
+func (i *RoleMappingInter) Find(filter *interfaces.Filter, ownerRelations []domain.Relation) ([]domain.RoleMapping, error) {
+	rolemappings, err := i.repo.Find(filter, ownerRelations)
 	return rolemappings, err
 }
 
-func (i *RoleMappingInter) FindByID(id int, filter *interfaces.Filter) (*domain.RoleMapping, error) {
-	rolemapping, err := i.repo.FindByID(id, filter)
+func (i *RoleMappingInter) FindByID(id int, filter *interfaces.Filter, ownerRelations []domain.Relation) (*domain.RoleMapping, error) {
+	rolemapping, err := i.repo.FindByID(id, filter, ownerRelations)
 	return rolemapping, err
 }
 
-func (i *RoleMappingInter) Upsert(rolemappings []domain.RoleMapping) ([]domain.RoleMapping, error) {
-	rolemappings, err := i.repo.Upsert(rolemappings)
+func (i *RoleMappingInter) Upsert(rolemappings []domain.RoleMapping, filter *interfaces.Filter, ownerRelations []domain.Relation) ([]domain.RoleMapping, error) {
+	rolemappings, err := i.repo.Upsert(rolemappings, filter, ownerRelations)
 	return rolemappings, err
 }
 
-func (i *RoleMappingInter) UpsertOne(rolemapping *domain.RoleMapping) (*domain.RoleMapping, error) {
-	rolemapping, err := i.repo.UpsertOne(rolemapping)
+func (i *RoleMappingInter) UpsertOne(rolemapping *domain.RoleMapping, filter *interfaces.Filter, ownerRelations []domain.Relation) (*domain.RoleMapping, error) {
+	rolemapping, err := i.repo.UpsertOne(rolemapping, filter, ownerRelations)
 	return rolemapping, err
 }
 
-func (i *RoleMappingInter) DeleteAll(filter *interfaces.Filter) error {
-	err := i.repo.DeleteAll(filter)
+func (i *RoleMappingInter) DeleteAll(filter *interfaces.Filter, ownerRelations []domain.Relation) error {
+	err := i.repo.DeleteAll(filter, ownerRelations)
 	return err
 }
 
-func (i *RoleMappingInter) DeleteByID(id int) error {
-	err := i.repo.DeleteByID(id)
+func (i *RoleMappingInter) DeleteByID(id int, filter *interfaces.Filter, ownerRelations []domain.Relation) error {
+	err := i.repo.DeleteByID(id, filter, ownerRelations)
 	return err
 }
