+++ ressources/templates/interactors/crud.go
@@ -32,12 +32,13 @@ var interactor = &typewriter.Template{
 	type Abstract{{.Type}}Repo interface {
 		Create({{.Name}}s []domain.{{.Type}}) ([]domain.{{.Type}}, error)
 		CreateOne({{.Name}} *domain.{{.Type}}) (*domain.{{.Type}}, error)
-		Find(filter *interfaces.Filter) ([]domain.{{.Type}}, error)
-		FindByID(id int, filter *interfaces.Filter) (*domain.{{.Type}}, error)
-		Upsert({{.Name}}s []domain.{{.Type}}) ([]domain.{{.Type}}, error)
-		UpsertOne({{.Name}} *domain.{{.Type}}) (*domain.{{.Type}}, error)
-		DeleteAll(filter *interfaces.Filter) error
-		DeleteByID(id int) error
+		Find(filter *interfaces.Filter, ownerRelations []domain.Relation) ([]domain.{{.Type}}, error)
+		FindByID(id int, filter *interfaces.Filter, ownerRelations []domain.Relation) (*domain.{{.Type}}, error)
+		Upsert({{.Name}}s []domain.{{.Type}}, filter *interfaces.Filter, ownerRelations []domain.Relation) ([]domain.{{.Type}}, error)
+		UpsertOne({{.Name}} *domain.{{.Type}}, filter *interfaces.Filter, ownerRelations []domain.Relation) (*domain.{{.Type}}, error)
+		DeleteAll(filter *interfaces.Filter, ownerRelations []domain.Relation) error
+		DeleteByID(id int, filter *interfaces.Filter, ownerRelations []domain.Relation) error
+		Raw(query string, values ...interface{}) (*sql.Rows, error)
 	}
 
 	type {{.Type}}Inter struct {
@@ -70,8 +71,8 @@ var createOne = &typewriter.Template{
 var find = &typewriter.Template{
 	Name: "Find",
 	Text: `
-	func (i *{{.Type}}Inter) Find(filter *interfaces.Filter) ([]domain.{{.Type}}, error) {
-		{{.Name}}s, err := i.repo.Find(filter)
+	func (i *{{.Type}}Inter) Find(filter *interfaces.Filter, ownerRelations []domain.Relation) ([]domain.{{.Type}}, error) {
+		{{.Name}}s, err := i.repo.Find(filter, ownerRelations)
 		return {{.Name}}s, err
 	}
 `}
@@ -79,8 +80,8 @@ var find = &typewriter.Template{
 var findByID = &typewriter.Template{
 	Name: "FindByID",
 	Text: `
-	func (i *{{.Type}}Inter) FindByID(id int, filter *interfaces.Filter) (*domain.{{.Type}}, error) {
-		{{.Name}}, err := i.repo.FindByID(id, filter)
+	func (i *{{.Type}}Inter) FindByID(id int, filter *interfaces.Filter, ownerRelations []domain.Relation) (*domain.{{.Type}}, error) {
+		{{.Name}}, err := i.repo.FindByID(id, filter, ownerRelations)
 		return {{.Name}}, err
 	}
 `}
@@ -88,8 +89,8 @@ var findByID = &typewriter.Template{
 var upsert = &typewriter.Template{
 	Name: "Upsert",
 	Text: `
-	func (i *{{.Type}}Inter) Upsert({{.Name}}s []domain.{{.Type}}) ([]domain.{{.Type}}, error) {
-		{{.Name}}s, err := i.repo.Upsert({{.Name}}s)
+	func (i *{{.Type}}Inter) Upsert({{.Name}}s []domain.{{.Type}}, filter *interfaces.Filter, ownerRelations []domain.Relation) ([]domain.{{.Type}}, error) {
+		{{.Name}}s, err := i.repo.Upsert({{.Name}}s, filter, ownerRelations)
 		return {{.Name}}s, err
 	}
 `}
@@ -97,8 +98,8 @@ var upsert = &typewriter.Template{
 var upsertOne = &typewriter.Template{
 	Name: "UpsertOne",
 	Text: `
-	func (i *{{.Type}}Inter) UpsertOne({{.Name}} *domain.{{.Type}}) (*domain.{{.Type}}, error) {
-		{{.Name}}, err := i.repo.UpsertOne({{.Name}})
+	func (i *{{.Type}}Inter) UpsertOne({{.Name}} *domain.{{.Type}}, filter *interfaces.Filter, ownerRelations []domain.Relation) (*domain.{{.Type}}, error) {
+		{{.Name}}, err := i.repo.UpsertOne({{.Name}}, filter, ownerRelations)
 		return {{.Name}}, err
 	}
 `}
@@ -106,8 +107,8 @@ var upsertOne = &typewriter.Template{
 var deleteAll = &typewriter.Template{
 	Name: "DeleteAll",
 	Text: `
-	func (i *{{.Type}}Inter) DeleteAll(filter *interfaces.Filter) error {
-		err := i.repo.DeleteAll(filter)
+	func (i *{{.Type}}Inter) DeleteAll(filter *interfaces.Filter, ownerRelations []domain.Relation) error {
+		err := i.repo.DeleteAll(filter, ownerRelations)
 		return err
 	}
 `}
@@ -115,8 +116,8 @@ var deleteAll = &typewriter.Template{
 var deleteByID = &typewriter.Template{
 	Name: "DeleteByID",
 	Text: `
-	func (i *{{.Type}}Inter) DeleteByID(id int) error {
-		err := i.repo.DeleteByID(id)
+	func (i *{{.Type}}Inter) DeleteByID(id int, filter *interfaces.Filter, ownerRelations []domain.Relation) error {
+		err := i.repo.DeleteByID(id, filter, ownerRelations)
 		return err
 	}
 `}
